[chore/protocol-testing] move test_lang to simulations/
